Switch to atoms.
This one's really big. This could be thought of as sort of a 2.0 moment. I
will *not* squash this commit.

Instead of passing around verbs and implying arity from length of argument
lists, consolidate them into atoms, which wrap verbs and arity and are easily
promotable by the JIT.

Additionally, improve several JIT annotations, including atom promotion, node
promotion, and immutability of certain lists. We're looking at traces that are
half the length of what they used to be.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>