nodes: The Node Identity.
First of the Node refactors.

This is the best slice of isolated changes that I could divine, and I spent a
good half-day of thought on the problem. Basically, nodes have changed their
internal layout, and that means that a lot of node-handling code needs to be
changed to work with them. In particular, since the members of nodes have
changed, their pretty-printing needs to change. But that means that we have to
pull in the newer AST loader and the pretty-printing helper.

I compromised here and gutted ScriptObject. For this commit (and only this
commit!) ScriptObjects do not work.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>