nodes: The Node Supremacy.
The short version: We needed to restore ScriptObject. From that, all else
followed.

Patterns now always use ejectors when unifying. This creates a circle between
ejection-as-control and patterns-as-definitions which ends up pulling some
stuff in.

This patch finally kills t.o.__init__, which is pretty cool. It also
introduces ScriptMaps, which are a much-needed structure for optimization.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>