Introduce list strategies.
List strategies are a technique for eliminating the extra indirection in
containers which contain boxed objects, by detecting patterns in the boxes and
removing them. Strategies should not hurt the quality of JIT traces, and
should actually improve the JIT's speed in situations where lists are
homogenous.

The actual impact of list strategies is as follows: Stuff gets a little
slower, code gets harder to read, and overall Corbin gets a little sadder.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>