SmallCaps.
This commit introduces a customized SmallCaps VM and an ad-hoc compiler from
Kernel-Monte to SmallCaps.

JIT translation is working; however, only local binding accesses are
virtualized. References to the outer frame are not. There's some other JIT
inefficiencies that haven't yet been corrected.

Only final patterns are supported for self-referential objects.

Not all userspace code is confirmed working, but most of it should be fine.
Mostly.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>