smallcaps: Improve JIT annotations.
Elides a large amount of work related to calls by proving that atoms from the
code object are green. Elides a lot of work in the rest of the opcodes, too;
repeated loads and stores are much better now.

Wondering whether a cheap little peephole bytecode optimizer would be a good
investment. It'd only take a day.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>