Use the same frame order for every method and matcher on an object.
The goal is to make one single closure per object and reuse it, rather than
create it anew for every call.

Additionally, this causes the closure to be a single virtualizable array that
is shared between different method invocations, which improves the ability of
the JIT to remove stuff. Many stack-manipulating opcodes are now nops in the
JIT trace.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>