smallcaps: Add a peephole optimizer.
It comes prepopulated with the most useful optimizations, removing the
extraneous POPs inserted by Sequences.

The first draft of this code had a bug where it would destroy else-clauses,
sometimes with incorrect results. This version will not remove else-clauses,
to avoid unbalancing the stack.

This optimization is relatively slow, since it can backtrack to find more and
more optimizations. It'll take linear time, but linear with large constants,
to find pathological optimization opportunities, but the results are almost
certainly worth it since code is optimized exactly once and never freed.

Massive cuts to nqueens, montstone, etc.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>