reactor: Handle SIGINT.
Of course, that's too short of a commit message for the number of lines
changed, so here's the longer version.

We add handling for all signals via the wakeup FD technique. This requires
that a wakeup FD be added to the reactor. Rather than specialize the reactor
on the wakeup FD, we instead generalize sockets to "selectables" (Twisted's
term, not mine) and create a new selectable for the wakeup FD.

The rest of the commit is shuffling, moving the RPython socket code away from
the socket founts and draings, and making the entire enchilada shut down
gracefully. (I gave up and just use an exception.)

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>