Make flowStopped/0 into flowStopped/1.
This is a departure from Twisted's tube model; we are passing an explicit
reason between tubes when a pipeline stops flow. The reason can be something
as simple as "end of input", and is meant to provide error handling or error
logging. Perhaps a NullOk[Str] convention will evolve over time.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>