typhon/objects/data: Add bigints.
Includes patches to boot scope to make Int work, and patches to equality
checking so that ints are correctly compared for equality regardless of
whether or not they are bigints. Also improves unwrappers to have a more
distinct error.

Some tests are included. I wasn't up to the task of typing out every single
possible path of execution. This'll surely bite me in the ass later.

This patch is, to say the least, not battle-tested. I will be unsurprised when
flaws are found with it over the next few days.

Signed-off-by: Corbin Simpson <cds@corbinsimpson.com> <simpsoco@google.com>