Split frames into global and non-global bindings.
The goal is to enable the JIT to treat globals (bindings which come from
outside the module) as greens instead of reds. The justification is that
compilation is done per-module and so the greens which are bound into the
module are extremely unlikely to change. In fact, any object coming from a
module must have a green key which is unique to that module, and modules are
usually bound once per compilation.

Currently this commit has compiler and machine changes necessary to handle
globals separately from the rest of the frame, and works untranslated.

Translation works, is correct, and continues to JIT as before. The expected
speedups are less-than-dramatic on benchmarks, but HTTP seems quite a bit
faster! siege reports roughly 2x (!?!) transactions/second and the JIT'd code
is definitely tighter than before. I think that some more stuff can be
squeezed out with relative ease, if I focus on it.

The disassembler can print the new _GLOBAL opcodes, so the logs should be just
as readable as before.