smallcaps/machine: Fix long-standing stack-mishandling bug.
Allen helped me narrow this one down.

When exiting scopes delimited by TRY and END_HANDLER, the handler installed by
TRY is dropped at END_HANDLER. The dropped handler should branch the machine
to the point after the catch-expr of the try-expr, but it wasn't.

This caused all kinds of trouble, and I suspect that this was the only reason
that the stack sometimes overflowed.

Translates fine.