monte_optimizer: Narrow escapes, and remove certain small escapes.
This commit is known to provoke weirdness in some switch trees. Not sure why
yet; current theory is that a latent off-by-one in t.nodes compiling escapes
with catches (which are not well-tested) is exacerbated by trimming the
contents of the escape body.

mast/lib/codec/utf8.mt is affected by this particular bug! It's okay, though;
this commit has been tweaked to work around it. The committed boot should be
correct.