Scope vats precisely to only be global during on-vat execution.
This commit contains enough legwork to make the HTTP server and IRC bot work
again after the breaking changes introduced during the vat rescope.

Vats now live in a scoped list of vats. This is obvious precursor work to
running two vats on-CPU simultaneously, but is also needed for the
implementation of Vat.seed/1, which is close enough to smell. (And taste.)