objects/slots: Specialize FinalSlots within Bindings.
The specialization is incomplete and only happens when statically known.
That's a fancy way of saying that I only changed the two spots in Typhon that
had `Binding(FinalSlot(...), ...)` to use FinalBinding instead.

The improvements are modest but extant. Benchmarks have savings of 5-10% in
memory usage and runtime. The heap appears a little smaller, by maybe ~120KiB
or so. This shrinkage isn't expected to scale, since most of the improvements
are on one-time data structures. There *should* be a very slight
per-user-defined-object improvement but I was unable to measure it.