monte_expander: Optimistically expand simple logical conditionals.
Anything of the form m`if (x && y): ...` or m`if (x || y): ...` is expanded to
duplicate either the consequent or alternative. For or-exprs, this only is
done as long as no name defined in m`x` is used in m`y`. This is a slight
increase in code size but it remains *much* smaller than the equivalent
heavy-duty expansion normally brought to bear on this idiom.