prelude: Provide a lazy slot primitive.
We have to provide it like this so that it can be a valid thing for DFness; we
couldn't convince the full DF auditor to allow these slots otherwise.

The guard being specified twice sucks, but the expansion of _slotToBinding
doesn't give us the guard. Or maybe it does and I just didn't see how it
worked; it's a little late here.