smallcaps: Switch to "zero-cost" exception unwinding.
Under this model, similar to modern C++ compilers, we precalculate the
unwinding positions for all exception handlers, and store them in a table. We
no longer have an exception-handler stack; instead, we iteratively visit each
handler until we rescue the evaluation context or throw to the next frame.

Our approach here forces us to commit to a precise ordering of exception
effects. I've supplied a basic and useful semantic behavior: Ejectors are
trumped by exceptions, and exceptions chain so that all exception information
is preserved. (Actual chaining behavior to be implemented in a later commit.)