metrics: Refactor to allow overlapping contexts.
We've had contexts incidentally overlapping for a while, but this change
breaks all overlaps up into segments, and only tallies one segment at a time.
This allows arbitrary overlapping of contexts with only a slight inefficiency.

In untranslated tests, the cost of metrics is 3% at startup, diminishing to 1%
over time. Translated, the cost of metrics is less than 0.01% at startup.