prelude/monte_ast: Fix pretty-printing of MapPatternAssoc.
Look closely at this commit; it's confusing. Here, a type annotation `:Int`
might be thought to find the bug, which is passing a `Str` to a comparison
with an `Int`. However, upon closer inspection, the bug is not reproducible in
isolation, and the type annotation is really an afterthought from the original
bug fix, to quickly confirm that no other instances of the bug are lurking.