objects/data: Fix isInt() to correspond precisely to unwrapInt().
I know that it's a bit of a waste when interpreted, but it's the only way to
be correct when translating on CPython.

Maybe an explanation is in order. We use the isX() functions to determine
whether we should enter a specific multimethod; once we've chosen the
multimethod, we commit and use unwrapX() to complete the dispatch. "Look
before you leap." However, this requires a perfect correspondence; if isX()
lies, then unwrapX() could raise.

Why this didn't come up before: Iterating over members of classes in PyPy is
deterministic due to ordered dicts, so I just put the BigInt versions before
the Int versions and it worked out.