mix: Put a cap on this particular BTA algorithm.
It's not bad, but I need to capture two quirks which I believe are necessary
for fast Monte:

* Speedy polyvariance: We need to detect when there are possible known values
  for a name in a case which is distinct from normal static values. This is
  because we need all of those values during specialization, so that we can
  duplicate the subordinate clauses exactly once per polyvariant name.
* Proper escape analysis: We need extremely correct escape analysis if we want
  to be aggressive about unfolding all static calls, especially if we want to
  unroll loops with VarSlots in them.

  We need to be able to only partially unroll loops, and VarSlots make that
  difficult. For example, let's take this example:

    def triangle(x :Int):
        var a :Int := 0
        for i in (0..x) { a += i }
        return a

  Here, `a` could escape the loop if `x` is bigger than the number of
  iterations that we are willing to unroll. This is because the for-loop is
  expanded to a call:

    _loop(0..x, fn _, i { a += i })

  Since this inner function closes over `a`, we must assume that `a` escapes
  if the function escapes, and the function escapes in the `_loop` call. (You
  might protest that `DeepFrozen` prevents it, but consider Map.fetch/2
  first!)

  A partial unroll, with perhaps three unrollings, would then look like:

    a += 0
    a += 1
    a += 2
    _loop([3, 4, 5], fn _, i { a += i })

  Note that we do not partially unroll the first parameter of `_loop` to the
  expected `3..5`, because Monte iterators usually are not uncallable. It
  would be delightful if they were, because then they would be usable as
  zippers; however, this isn't a consideration right now. Instead, we can
  unroll finite iterators if we can show them to be statically finite, and we
  can unroll `_iterForever` a bounded number of times.

  Our reducer technology isn't to the point of partial loop unrolling, but my
  experiments here have shown total loop unrolling in the case of the
  Brainfuck interpreter example, and so I'm inclined to believe that we can
  get there with this design eventually. The annotator is what must improve.

How did we fail here? Well, in both of these cases, the underlying fault is
the same: We write down an annotation once when we bind a name, and then we
alter the annotation later when we observe the actual polyvariance or escape.
Unfortunately, by then, it is *too late* and, even if we patch the bind
location and the poly/escape location, we have already written the assumptions
down in intervening frames and cannot easily backtrack. The ad-hoc logic I've
tried to write just isn't correct.

We must instead seek a fixed point by starting with the strongest assumption,
which is that all names are static, and then finding contradictions to that
assumption. We use the contradictions to craft a weaker assumption, and
iterate until we find a fixed point assumption which is precisely the
static/dynamic annotations that we want.

Squash all the unpushed commits up to here. The history just isn't
interesting; it's broken all over.