fun/monads: Use sequence().
This supremely-useful combinator actually runs in O(n**2), but I'm willing to
live with that for now. To lower that, we need a more efficient list-building
operation. Maybe we could run this impurely, but that's gonna suck for when
anybody wants accurate monad effects, and that was kind of the entire point of
this monad exercise.