lib/uKanren: Buff and make working again.
The monad was a little screwy and assumed Scheme-style cons rather than
Monte-style pairs.

We weren't really delaying at all, so we had a very eager search system, only
delayed by fresh vars. Now we're delaying only when we have a chance to
branch. Note that unification is not just immediate, but also can prune
branches, so it is actually desirable to force unification goals!