rats: Implement rounding.
Rounding is a little different for rats. You can round to within a certain
epsilon, or you can truncate the amount of storage being used to store your
number. In both cases, it is possible for the rat to know that it might be
smaller and more accurate to stay unrounded.