interp: Extremely basic staged specializer.
♩ Well, I like to dream ♪ Yeah, right between my sound machine ♪

What's going on here? Well, for a given AST fragment and a context in which to
interpret it, we compute in big steps a simpler but equivalent fragment. By
storing optimizing information in the context, we can remove large amounts of
code directly.

I'm sure that I'll find at least one surprising incorrectly-handled-ejector
bug, but that is in the future and not the right-now.