objects/processes: Empty out os.environ (and thus empty out envp)
This change is a little sweeping, but the idea is easy: We want PATH, and
anything else in envp, to be blanked out so that glibc can't sneakily use it
as a covert channel. We don't make a special effort to do this early; we do it
when we construct the unsafe scope. We could do it earlier, but we need to
make sure to pickpocket os.environ first for currentProcess.

Seems to work, but we don't really have tests for this kind of process
integration.

Refs #155. This is a stepping stone towards that, and in any case it is
desirable to rob glibc of any interesting ambient authority.