games/tinyraytracer: The rest of the reasonable stuff.
The final commit in the tutorial requires C++'s behavior in the extreme,
and I'm not yet sure how to replicate it. Honestly, I might rather
produce a proper raytracer instead, but that would require admitting
that I'm taking this all rather seriously.

There is the unfortunate fact that this module's rendered output looks
only barely like the tutorial. I'm not sure how to rectify this; I've
never written a raytracer by transcribing a tutorial in C++ before,
although the tutorial assured me that it was a reasonable thing to
attempt.