objects/data: Improve and rename Double methods.
The renamed methods are not much used, so I've included fixups for all
of the corresponding Monte code.

The square root and logarithm edge-cases for signed zero are based on
pretty basic principles, but go against the grain of Python, and so
required effort. A key helper is the isNegative() test.

I have omitted atan(). All arctangent calls are going to be
two-argument, so that it's impossible to mess up the sign-to-angle
logic.

The Euclidean norm method is better than the corresponding open-coded
Monte. It's not just faster and harder to fuck up, but also more
accurate for extreme inputs.