lib/console: Add supersampling.
This seems like entirely the wrong layer at which to add on
supersampling, since the pixel density is key to the whole operation.

This is because the window in which to do supersampling, in terms of
distance, is extremely close to the pixel being sampled. If the window
gets even a little too large, then the supersampling not only falls off
and cancels out, but it starts to twist and distort. But if the window
is too small, which is really easy as well, then the effectiveness of
the technique drops off rapidly.