Centralize color handling.
A lot of this color-handling logic is actually brand new to our
codebase, but it's worth figuring out in one single place.

I will try to avoid not using codegen for the Porter-Duff algebra. I
think that there's a nice way to do it, but I need to think on it a
little more.