games/animate: Composite multiple triangles.
Yikes, this is slow. I will have to go back down into compiler guts to
get this to be fast.

It all works! This is definitely more LOC than it should be for just a
demo, but that's okay, because the parts that are generic are going to
be useful for building REPL tooling later.