games/sdf: Split out CSG -> SDF compilation.
This is sort of a demonstration of the power of casually using our ASDL
infrastructure. While this is the sort of situation where I'd love to
show off something like staged compilation, the fact is that even this
naïve compilation is still worth almost a 2x speedup in actual px/s.