Factor discrete sampling, supersampling, etc.
New module lib/samplers holds both the various sampler routines, as well
as ASDL for specifying a sampler by using abstract syntax for
configuration.

I've tested a bunch of these already. Some of them are very
time-consuming, and I'm not sure how to properly document that. Like, a
quincunx sample is 5x the cost of a center sample, while a t-test can
explode in consumed time. But I suppose that this is also the nature of
rendering; one has to know what they are asking for, and what it will
likely cost them in runtime.