lib/muffin: Share modules and module parameters.
This is not as radical of a rewrite as it seems; this mostly brings
lib/muffin back into harmony with loader.mt's recent changes. But also,
module parameters will be passed into each subordinate module; this
gives a way to parameterize an entire application by paramterizing each
module independently and then instantiating the corresponding muffin.